+++ affinda/models/_models.py
@@ -14974,7 +14974,7 @@ class ResthookSubscription(msrest.serialization.Model):
      "invoice.parse.succeeded", "invoice.parse.failed", "invoice.parse.completed",
      "invoice.validate.completed", "document.parse.succeeded", "document.parse.failed",
      "document.parse.completed", "document.validate.completed", "document.classify.succeeded",
-     "document.classify.failed", "document.classify.completed".
+     "document.classify.failed", "document.classify.completed", "document.rejected".
     :vartype event: str or ~affinda.models.ResthookEvent
     :ivar organization: Required.
     :vartype organization: ~affinda.models.Organization
@@ -15028,7 +15028,7 @@ class ResthookSubscription(msrest.serialization.Model):
          "invoice.parse.succeeded", "invoice.parse.failed", "invoice.parse.completed",
          "invoice.validate.completed", "document.parse.succeeded", "document.parse.failed",
          "document.parse.completed", "document.validate.completed", "document.classify.succeeded",
-         "document.classify.failed", "document.classify.completed".
+         "document.classify.failed", "document.classify.completed", "document.rejected".
         :paramtype event: str or ~affinda.models.ResthookEvent
         :keyword organization: Required.
         :paramtype organization: ~affinda.models.Organization
@@ -15072,7 +15072,7 @@ class ResthookSubscriptionCreate(msrest.serialization.Model):
      "invoice.parse.succeeded", "invoice.parse.failed", "invoice.parse.completed",
      "invoice.validate.completed", "document.parse.succeeded", "document.parse.failed",
      "document.parse.completed", "document.validate.completed", "document.classify.succeeded",
-     "document.classify.failed", "document.classify.completed".
+     "document.classify.failed", "document.classify.completed", "document.rejected".
     :vartype event: str or ~affinda.models.ResthookEvent
     :ivar organization:
     :vartype organization: str
@@ -15105,7 +15105,7 @@ class ResthookSubscriptionCreate(msrest.serialization.Model):
          "invoice.parse.succeeded", "invoice.parse.failed", "invoice.parse.completed",
          "invoice.validate.completed", "document.parse.succeeded", "document.parse.failed",
          "document.parse.completed", "document.validate.completed", "document.classify.succeeded",
-         "document.classify.failed", "document.classify.completed".
+         "document.classify.failed", "document.classify.completed", "document.rejected".
         :paramtype event: str or ~affinda.models.ResthookEvent
         :keyword organization:
         :paramtype organization: str
@@ -15131,7 +15131,7 @@ class ResthookSubscriptionUpdate(msrest.serialization.Model):
      "invoice.parse.failed", "invoice.parse.completed", "invoice.validate.completed",
      "document.parse.succeeded", "document.parse.failed", "document.parse.completed",
      "document.validate.completed", "document.classify.succeeded", "document.classify.failed",
-     "document.classify.completed".
+     "document.classify.completed", "document.rejected".
     :vartype event: str or ~affinda.models.ResthookEvent
     :ivar organization: Uniquely identify an organization.
     :vartype organization: str
@@ -15156,7 +15156,7 @@ class ResthookSubscriptionUpdate(msrest.serialization.Model):
          "invoice.parse.failed", "invoice.parse.completed", "invoice.validate.completed",
          "document.parse.succeeded", "document.parse.failed", "document.parse.completed",
          "document.validate.completed", "document.classify.succeeded", "document.classify.failed",
-         "document.classify.completed".
+         "document.classify.completed", "document.rejected".
         :paramtype event: str or ~affinda.models.ResthookEvent
         :keyword organization: Uniquely identify an organization.
         :paramtype organization: str