+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Allow creating workspace-scope webhook
+- Add "document.rejected" webhook event
 
 ## [4.9.0] - 2023-07-26
 ### Added